* unify add to recent hooks
